+++ CHANGELOG.md
@@ -23,6 +23,7 @@ type Foo = {
 
 - Sync to upstream Luau 0.604
 - Overhauled command line argument parsing system to be more consistent and flexible
+- Deprioritized `loadstring` in autocomplete
 
 ### Fixed